feat: update rpms and images data fetching
Refactor code to improve efficiency and readability
Replace requests dependency with ghapi
Update data fetching logic to match new API usage
